added search view, fixed base url

+++ app/src/main/java/com/example/achyp/weatherapp/pojo/MainInfo.java
@@ -2,15 +2,12 @@ package com.example.achyp.weatherapp.pojo;
 
 import com.google.gson.annotations.SerializedName;
 
-/**
- * Created by achyp on 2/18/17.
- */
 
 public class MainInfo {
     @SerializedName("temp")
     public float temperature;
-    public int pressure;
-    public int humidity;
+    public float pressure;
+    public float humidity;
     public float temp_min;
     public float temp_max;
 }
