+++ app/src/main/java/com/example/achyp/weatherapp/pojo/Wind.java
@@ -1,10 +1,6 @@
 package com.example.achyp.weatherapp.pojo;
 
-/**
- * Created by achyp on 2/18/17.
- */
-
 public class Wind {
     public float speed;
-    public int deg;
+    public float deg;
 }
